scripts/requirements.txt
Pillow>=10.0.0
pdf2image>=1.16.0
pypdf>=3.0.0
reportlab>=4.0.0

